Revert Move Bootstrap.class to presto-bytecode to avoid duplicate definitions [D75970190] (#25260)
Summary:

This diff reverts D75970190
Reverts https://github.com/prestodb/presto/pull/25245?fbclid=IwY2xjawKwI6BleHRuA2FlbQIxMQBicmlkETBoYW9yclAzdjR2bW9rU1ZlAR5T6z9TdGaufXbSNH8Y7s5HI1NhghLSwgqbKWZLniAvsDvnZshdKowS2qj08A_aem_CLbW4rp6V1haCMjCgdqPDg
Causing issues in our facebook internal verifier. Not sure if issue is with facebook internal, or oss spi. Temporary revert so we can continue with release while we attempt to fix the issue.

Depends on D75970190

Differential Revision: D76145724
